sct_viewer: lazyloading applied
Former-commit-id: fe2b439bd2de513c8b9e373e9db309d83bf5a927 [formerly 878e0ec2838e4505af02292ab1adc5b5fa86dc1e [formerly cf12576c77e9cd49bd42de1b3f5667929e752eb8]]
Former-commit-id: 29ac94a077b0df83ec3f41ec0bbd52e6cef5afea
Former-commit-id: b45d2304c947e8e0b0286b469304a447078ff872
